Move keyboard UI utilities
